refactor: signup type 관리
